move datasetsampleactions.js
